+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Fixed a wrong JNI method declaration which might cause "method not found" crash on some devices.
 * Fixed a bug that `Error` in the background async thread is not forwared to the caller thread.
 * Fixed a crash when an empty `Collection` is passed to `insert()`/`insertOrUpdate()` (#3103).
-* Fixed a bug that does not transfer a primary key when `RealmSchemaObject.setClassName()` is called to rename a class (#3118).
+* Fixed a bug that does not transfer the primary key when `RealmSchemaObject.setClassName()` is called to rename a class (#3118).
 
 ## 1.1.0